// sql/ast/interfaces/ISqlNode.d.ts
export default interface ISqlNode {
    _location?: INodeLocation;
}
// ...
export interface INodeLocation {
    /** Location of the last ";" prior to this statement */
    start: number;
    /** Location of the first ";" after this statement (if any) */
    end: number;
}
// ...
export interface IName extends ISqlNode {
    name: string;
}
export interface ITableAliasName extends IName, ISqlNode {
    columns?: IName[];
}
export interface IQName extends IName, ISqlNode {
    schema?: string;
}
export interface IQColumn extends ISqlNode {
    table: string;
    column: string;
    schema?: string;
}
export declare type IDataTypeDef = IArrayDataTypeDef | IBasicDataTypeDef;
export interface IArrayDataTypeDef extends ISqlNode {
    kind: 'array';
    arrayOf: IDataTypeDef;
}
export interface IBasicDataTypeDef extends IQName, ISqlNode {
    kind?: undefined;
    /** Allows to differenciate types like 'double precision' from their double-quoted counterparts */
    doubleQuoted?: true;
    /** varchar(length), numeric(precision, scale), ... */
    config?: number[];
}
export interface ITableReference {
    foreignTable: IQName;
    foreignColumns: IName[];
    match?: 'full' | 'partial' | 'simple';
}
export declare type ISelectStatement = ISelectFromStatement | IValuesStatement;
export interface ISelectFromStatement extends ISqlNode {
    type: 'select';
    columns?: ISelectedColumn[] | INil;
    from?: IFrom[] | INil;
    where?: IExpr | INil;
    groupBy?: IExpr[] | INil;
    limit?: ILimitStatement | INil;
    orderBy?: IOrderByStatement[] | INil;
    distinct?: 'all' | 'distinct' | IExpr[] | INil;
    for?: IForStatement;
}
export interface IOrderByStatement extends ISqlNode {
    by: IExpr;
    order?: 'ASC' | 'DESC' | INil;
    nulls?: 'FIRST' | 'LAST' | INil;
}
// ...
export interface IJoinClause extends ISqlNode {
    type: IJoinType;
    on?: IExpr | INil;
    using?: IName[] | INil;
}
export declare type IJoinType = 'INNER JOIN' | 'LEFT JOIN' | 'RIGHT JOIN' | 'FULL JOIN' | 'CROSS JOIN';
export declare type IExpr = IExprRef | IExprParameter | IExprList | IExprArrayFromSelect | IExprNull | IExprExtract | IExprInteger | IExprDefault | IExprMember | IExprValueKeyword | IExprArrayIndex | IExprNumeric | IExprString | IExprCase | IExprBinary | IExprUnary | IExprCast | IExprBool | IExprCall | ISelectStatement | IExprConstant | IExprTernary | IExprOverlay | IExprSubstring;
/**
 * Handle special syntax: overlay('12345678' placing 'ab' from 2 for 4)
 */
export interface IExprOverlay extends ISqlNode {
    type: 'overlay';
    value: IExpr;
    placing: IExpr;
    from: IExpr;
    for?: IExpr | INil;
}
/** Handle special syntax: substring('val' from 2 for 3) */
export interface IExprSubstring extends ISqlNode {
    type: 'substring';
    value: IExpr;
    from?: IExpr | INil;
    for?: IExpr | INil;
}
// ...
export interface IExprBinary extends ISqlNode {
    type: 'binary';
    left: IExpr;
    right: IExpr;
    op: IBinaryOperator;
    opSchema?: string;
}
export interface IExprConstant extends ISqlNode {
    type: 'constant';
    dataType: IDataTypeDef;
    value: any;
}
export declare type IExprLiteral = IExprConstant | IExprInteger | IExprNumeric | IExprString | IExprBool | IExprNull;
export interface IExprTernary extends ISqlNode {
    type: 'ternary';
    value: IExpr;
    lo: IExpr;
    hi: IExpr;
    op: 'BETWEEN' | 'NOT BETWEEN';
}
export interface IExprCast extends ISqlNode {
    type: 'cast';
    to: IDataTypeDef;
    operand: IExpr;
}
export declare type IUnaryOperator = '+' | '-' | 'NOT' | 'IS NULL' | 'IS NOT NULL' | 'IS TRUE' | 'IS FALSE' | 'IS NOT TRUE' | 'IS NOT FALSE';
export interface IExprUnary extends ISqlNode {
    type: 'unary';
    operand: IExpr;
    op: IUnaryOperator;
    opSchema?: string;
}
export interface IExprRef extends ISqlNode {
    type: 'ref';
    table?: IQName;
    name: string | '*';
}
export interface IExprParameter extends ISqlNode {
    type: 'parameter';
    name: string;
}
export interface IExprMember extends ISqlNode {
    type: 'member';
    operand: IExpr;
    op: '->' | '->>';
    member: string | number;
}
export interface IExprValueKeyword extends ISqlNode {
    type: 'keyword';
    keyword: IValueKeyword;
}
export declare type IValueKeyword = 'current_catalog' | 'current_date' | 'current_role' | 'current_schema' | 'current_timestamp' | 'current_time' | 'localtimestamp' | 'localtime' | 'session_user' | 'user' | 'current_user' | 'distinct';
/**
 * Function calls.
 *
 * For aggregation functions, see https://www.postgresql.org/docs/13/sql-expressions.html#SYNTAX-AGGREGATES
 */
export interface IExprCall extends ISqlNode {
    type: 'call';
    /** Function name */
    function: IQName;
    /** Arguments list */
    args: IExpr[];
    /** [AGGREGATION FUNCTIONS] Distinct clause specified ? */
    distinct?: 'all' | 'distinct';
    /** [AGGREGATION FUNCTIONS] Inner order by clause */
    orderBy?: IOrderByStatement[] | INil;
    /** [AGGREGATION FUNCTIONS] Filter clause */
    filter?: IExpr | INil;
    /** [AGGREGATION FUNCTIONS] OVER clause */
    over?: ICallOver | INil;
}
export interface ICallOver extends ISqlNode {
    orderBy?: IOrderByStatement[] | INil;
    partitionBy?: IExpr[] | INil;
}
export interface IExprExtract extends ISqlNode {
    type: 'extract';
    field: IName;
    from: IExpr;
}
export interface IExprList extends ISqlNode {
    type: 'list' | 'array';
    expressions: IExpr[];
}
export interface IExprArrayFromSelect extends ISqlNode {
    type: 'array select';
    select: ISelectStatement;
}
export interface IExprArrayIndex extends ISqlNode {
    type: 'arrayIndex';
    array: IExpr;
    index: IExpr;
}
export interface IExprNull extends ISqlNode {
    type: 'null';
}
export interface IExprInteger extends ISqlNode {
    type: 'integer';
    value: number;
}
export interface IExprDefault extends ISqlNode {
    type: 'default';
}
export interface IExprNumeric extends ISqlNode {
    type: 'numeric';
    value: number;
}
export interface IExprString extends ISqlNode {
    type: 'string';
    value: string;
}
export interface IExprBool extends ISqlNode {
    type: 'boolean';
    value: boolean;
}
export interface IExprCase extends ISqlNode {
    type: 'case';
    value?: IExpr | INil;
    whens: IExprWhen[];
    else?: IExpr | INil;
}